Extract PackageFilter class for clarity
